requirements.txt
ansicon==1.89.0
blessed==1.20.0
click==8.1.7
colorama==0.4.6
editor==1.6.6
inquirer==3.2.4
jinxed==1.2.1
pydobot==1.3.2
pyserial==3.4
readchar==4.0.5
runs==1.2.2
six==1.16.0
termcolor==2.4.0
typer==0.9.0
typing_extensions==4.10.0
wcwidth==0.2.13
xmod==1.8.1
yaspin==3.0.1
